refactor: remove utils barrel export in utils
