+++ src/commands/allCap/index.ts
@@ -1,7 +1,8 @@
 import { type ChatInputCommandInteraction, SlashCommandBuilder, type TextChannel } from 'discord.js';
 import { Result } from 'oxide.ts';
-import { fetchLastMessageBeforeId, isBlank } from '../../utils';
+import { isBlank } from '../../utils/isBlank';
 import { logger } from '../../utils/logger';
+import { fetchLastMessageBeforeId } from '../../utils/messageFetcher';
 import type { Command } from '../builder';
 
 const data = new SlashCommandBuilder()
