+++ src/commands/cowsay/index.ts
@@ -1,8 +1,9 @@
 import { say } from 'cowsay';
 import { type ChatInputCommandInteraction, SlashCommandBuilder, type TextChannel } from 'discord.js';
 import { Result } from 'oxide.ts';
-import { fetchLastMessageBeforeId, isBlank } from '../../utils';
+import { isBlank } from '../../utils/isBlank';
 import { logger } from '../../utils/logger';
+import { fetchLastMessageBeforeId } from '../../utils/messageFetcher';
 import type { Command } from '../builder';
 
 // Only 35 characters per line due to limitation in phone screen width
