+++ src/commands/referral/referralRandom.ts
@@ -1,7 +1,7 @@
 import { type Guild, SlashCommandSubcommandBuilder } from 'discord.js';
 import { Result } from 'oxide.ts';
-import { getRandomIntInclusive } from '../../utils';
 import { logger } from '../../utils/logger';
+import { getRandomIntInclusive } from '../../utils/random';
 import type { AutocompleteHandler, CommandHandler } from '../builder';
 import { getAllReferralCodesForService } from './referralUtils';
 import { searchServices } from './services';
